requirements.txt
pyscreenshot
Pillow
pytesseract
